parse_commit_date(): advance the scan pointer for the caller
We will update parse_commit_buffer() to not to stop at author/committer
lines in the later patches, so change the "buf" parameter of this function
to be the caller's scan pointer and update it before successful return.

While at it, give a bit more comments on what it is doing.

Signed-off-by: Junio C Hamano <gitster@pobox.com>