t9350: point out that refs are not updated correctly
The fast-export command writes no output when passed an empty range,
e.g. A..A.  This happens only when the corresponding commits are not
exported in the current fast-export run. This can happen either when the
relevant commit is already marked, or when the commit is explicitly marked
as UNINTERESTING with a negative ref by another argument.

This breaks fast-export based remote helpers, as they use marks
files to store which commits have already been seen. The call graph
is something as follows:

  # push master to remote repo
  $ git fast-export --{im,ex}port-marks=marksfile master
  # make a commit on master and push it to remote
  $ git fast-export --{im,ex}port-marks=marksfile master
  # run `git branch foo` and push it to remote
  $ git fast-export --{im,ex}port-marks=marksfile foo

When fast-export imports the marksfile and sees that all commits in
foo are marked as UNINTERESTING (they have already been exported
while pushing master), it exits without doing anything. However,
what we want is for it to reset 'foo' to the already-exported commit.

Either way demonstrates the problem, and since this is the most
succint way to demonstrate the problem it is implemented by passing
master..master on the commandline.

Signed-off-by: Sverre Rabbelier <srabbelier@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>