credential: use git_prompt instead of git_getpass
When we request a username and password from the user on the
terminal, we can't use stdin/stdout, because they may be
connected to pipes to other git processes.

Instead, we use getpass() (via git_getpass), which will
generally open /dev/tty and read and write from that.  The
only problem is that getpass is meant to take passwords, and
therefore will not echo characters from the username, which
is annoying.

Now that git_prompt understand the "echo" flag, we can use
that to let the user see their username as they type it.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>